Delay imports and force registration off the import path.
As title.
