+++ numba/core/cpu.py
@@ -16,7 +16,6 @@ from numba.core.compiler_lock import global_compiler_lock
 import numba.core.entrypoints
 from numba.core.cpu_options import (ParallelOptions, FastMathOptions,
                                     InlineOptions)
-from numba.cpython import setobj, listobj
 from numba.np import ufunc_db
 
 # Keep those structures in sync with _dynfunc.c.
@@ -81,6 +80,7 @@ class CPUContext(BaseContext):
         from numba.core import optional
         from numba.misc import gdb_hook, literal
         from numba.np import linalg, polynomial, arraymath
+        from numba.typed import typeddict
 
         try:
             from numba.np import npdatetime
@@ -153,12 +153,14 @@ class CPUContext(BaseContext):
         """
         Build a list from the Numba *list_type* and its initial *items*.
         """
+        from numba.cpython import listobj
         return listobj.build_list(self, builder, list_type, items)
 
     def build_set(self, builder, set_type, items):
         """
         Build a set from the Numba *set_type* and its initial *items*.
         """
+        from numba.cpython import setobj
         return setobj.build_set(self, builder, set_type, items)
 
     def build_map(self, builder, dict_type, item_types, items):
