+++ numba/typed/typedlist.py
@@ -8,13 +8,11 @@ and uses `@jit` functions to access it. Since it inherits from MutableSequence
 it should really quack like the CPython `list`.
 
 """
-from collections.abc import MutableSequence
 
 from numba.core.types import ListType, TypeRef
 from numba.core.imputils import numba_typeref_ctor
-from numba.core.dispatcher import Dispatcher
-from numba.core import types, config, cgutils
-from numba import njit, typeof
+from numba.core import types, cgutils
+from numba import njit
 from numba.core.extending import (
     overload_method,
     overload,
@@ -23,15 +21,12 @@ from numba.core.extending import (
     NativeValue,
     type_callable,
 )
-from numba.typed import listobject
+from numba.typed import api, listobject
+from numba.typed.api import List
 from numba.core.errors import TypingError, LoweringError
 from numba.core.typing.templates import Signature
 import typing as pt
 import sys
-
-
-Int_or_Slice = pt.Union["pt.SupportsIndex", slice]
-
 
 if sys.version_info >= (3, 8):
     T_co = pt.TypeVar('T_co', covariant=True)
@@ -44,7 +39,7 @@ if sys.version_info >= (3, 8):
             ...
 
 
-DEFAULT_ALLOCATED = listobject.DEFAULT_ALLOCATED
+DEFAULT_ALLOCATED = api.DEFAULT_ALLOCATED
 
 
 @njit
@@ -189,258 +184,6 @@ def _from_meminfo_ptr(ptr, listtype):
 
 T = pt.TypeVar('T')
 T_or_ListT = pt.Union[T, 'List[T]']
-
-
-class List(MutableSequence, pt.Generic[T]):
-    """A typed-list usable in Numba compiled functions.
-
-    Implements the MutableSequence interface.
-    """
-
-    _legal_kwargs = ["lsttype", "meminfo", "allocated"]
-
-    def __new__(cls,
-                lsttype=None,
-                meminfo=None,
-                allocated=DEFAULT_ALLOCATED,
-                **kwargs):
-        if config.DISABLE_JIT:
-            return list.__new__(list)
-        else:
-            return object.__new__(cls)
-
-    @classmethod
-    def empty_list(cls, item_type, allocated=DEFAULT_ALLOCATED):
-        """Create a new empty List.
-
-        Parameters
-        ----------
-        item_type: Numba type
-            type of the list item.
-        allocated: int
-            number of items to pre-allocate
-        """
-        if config.DISABLE_JIT:
-            return list()
-        else:
-            return cls(lsttype=ListType(item_type), allocated=allocated)
-
-    def __init__(self, *args, **kwargs):
-        """
-        For users, the constructor does not take any parameters.
-        The keyword arguments are for internal use only.
-
-        Parameters
-        ----------
-        args: iterable
-            The iterable to intialize the list from
-        lsttype : numba.core.types.ListType; keyword-only
-            Used internally for the list type.
-        meminfo : MemInfo; keyword-only
-            Used internally to pass the MemInfo object when boxing.
-        allocated: int; keyword-only
-            Used internally to pre-allocate space for items
-        """
-        illegal_kwargs = any((kw not in self._legal_kwargs for kw in kwargs))
-        if illegal_kwargs or args and kwargs:
-            raise TypeError("List() takes no keyword arguments")
-        if kwargs:
-            self._list_type, self._opaque = self._parse_arg(**kwargs)
-        else:
-            self._list_type = None
-            if args:
-                if not 0 <= len(args) <= 1:
-                    raise TypeError(
-                        "List() expected at most 1 argument, got {}"
-                        .format(len(args))
-                    )
-                iterable = args[0]
-                # Special case Numpy scalars or anything that quacks like a
-                # NumPy Array.
-                if hasattr(iterable, "ndim") and iterable.ndim == 0:
-                    self.append(iterable.item())
-                else:
-                    try:
-                        iter(iterable)
-                    except TypeError:
-                        raise TypeError("List() argument must be iterable")
-                    for i in args[0]:
-                        self.append(i)
-
-    def _parse_arg(self, lsttype, meminfo=None, allocated=DEFAULT_ALLOCATED):
-        if not isinstance(lsttype, ListType):
-            raise TypeError('*lsttype* must be a ListType')
-
-        if meminfo is not None:
-            opaque = meminfo
-        else:
-            opaque = _make_list(lsttype.item_type, allocated=allocated)
-        return lsttype, opaque
-
-    @property
-    def _numba_type_(self):
-        if self._list_type is None:
-            raise TypeError("invalid operation on untyped list")
-        return self._list_type
-
-    @property
-    def _typed(self):
-        """Returns True if the list is typed.
-        """
-        return self._list_type is not None
-
-    @property
-    def _dtype(self):
-        if not self._typed:
-            raise RuntimeError("invalid operation on untyped list")
-        return self._list_type.dtype
-
-    def _initialise_list(self, item):
-        lsttype = types.ListType(typeof(item))
-        self._list_type, self._opaque = self._parse_arg(lsttype)
-
-    def __len__(self) -> int:
-        if not self._typed:
-            return 0
-        else:
-            return _length(self)
-
-    def _allocated(self):
-        if not self._typed:
-            return DEFAULT_ALLOCATED
-        else:
-            return _allocated(self)
-
-    def _is_mutable(self):
-        return _is_mutable(self)
-
-    def _make_mutable(self):
-        return _make_mutable(self)
-
-    def _make_immutable(self):
-        return _make_immutable(self)
-
-    def __eq__(self, other):
-        return _eq(self, other)
-
-    def __ne__(self, other):
-        return _ne(self, other)
-
-    def __lt__(self, other):
-        return _lt(self, other)
-
-    def __le__(self, other):
-        return _le(self, other)
-
-    def __gt__(self, other):
-        return _gt(self, other)
-
-    def __ge__(self, other):
-        return _ge(self, other)
-
-    def append(self, item: T) -> None:
-        if not self._typed:
-            self._initialise_list(item)
-        _append(self, item)
-
-    # noqa F811 comments required due to github.com/PyCQA/pyflakes/issues/592
-    # noqa E704 required to follow overload style of using ... in the same line
-    @pt.overload  # type: ignore[override]
-    def __setitem__(self, i: int, o: T) -> None: ...  # noqa: F811, E704
-    @pt.overload
-    def __setitem__(self, s: slice, o: 'List[T]') -> None: ...  # noqa: F811, E704, E501
-
-    def __setitem__(self, i: Int_or_Slice, item: T_or_ListT) -> None:  # noqa: F811, E501
-        if not self._typed:
-            self._initialise_list(item)
-        _setitem(self, i, item)
-
-    # noqa F811 comments required due to github.com/PyCQA/pyflakes/issues/592
-    # noqa E704 required to follow overload style of using ... in the same line
-    @pt.overload
-    def __getitem__(self, i: int) -> T: ...  # noqa: F811, E704
-    @pt.overload
-    def __getitem__(self, i: slice) -> 'List[T]': ...  # noqa: F811, E704
-
-    def __getitem__(self, i: Int_or_Slice) -> T_or_ListT:  # noqa: F811
-        if not self._typed:
-            raise IndexError
-        else:
-            return _getitem(self, i)
-
-    def __iter__(self) -> pt.Iterator[T]:
-        for i in range(len(self)):
-            yield self[i]
-
-    def __contains__(self, item: T) -> bool:  # type: ignore[override]
-        return _contains(self, item)
-
-    def __delitem__(self, i: Int_or_Slice) -> None:
-        _delitem(self, i)
-
-    def insert(self, i: int, item: T) -> None:
-        if not self._typed:
-            self._initialise_list(item)
-        _insert(self, i, item)
-
-    def count(self, item: T) -> int:
-        return _count(self, item)
-
-    def pop(self, i: "pt.SupportsIndex" = -1) -> T:
-        return _pop(self, i)
-
-    def extend(self, iterable: "_Sequence[T]") -> None: #type: ignore[override]
-        # Empty iterable, do nothing
-        if len(iterable) == 0:
-            return None
-        if not self._typed:
-            # Need to get the first element of the iterable to initialise the
-            # type of the list. FIXME: this may be a problem if the iterable
-            # can not be sliced.
-            self._initialise_list(iterable[0])
-        return _extend(self, iterable)
-
-    def remove(self, item: T) -> None:
-        return _remove(self, item)
-
-    def clear(self):
-        return _clear(self)
-
-    def reverse(self):
-        return _reverse(self)
-
-    def copy(self):
-        return _copy(self)
-
-    def index(self, item: T, start: pt.Optional[int] = None,
-              stop: pt.Optional[int] = None) -> int:
-        return _index(self, item, start, stop)
-
-    def sort(self, key=None, reverse=False):
-        """Sort the list inplace.
-
-        See also ``list.sort()``
-        """
-        # If key is not already a dispatcher object, make it so
-        if callable(key) and not isinstance(key, Dispatcher):
-            key = njit(key)
-        return _sort(self, key, reverse)
-
-    def __str__(self):
-        buf = []
-        for x in self:
-            buf.append("{}".format(x))
-        # Check whether the code was invoked from IPython shell
-        try:
-            get_ipython
-            return '[{0}, ...]'.format(', '.join(buf[:1000]))
-        except (NameError, IndexError):
-            return '[{0}]'.format(', '.join(buf))
-
-    def __repr__(self):
-        body = str(self)
-        prefix = str(self._list_type) if self._typed else "ListType[Undefined]"
-        return "{prefix}({body})".format(prefix=prefix, body=body)
 
 
 # XXX: should we have a better way to classmethod
